# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Lost & Found API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ItemType = typeof ItemType[keyof typeof ItemType];


export const ItemType = {
  lost: 'lost',
  found: 'found',
} as const;

export type ItemStatus = typeof ItemStatus[keyof typeof ItemStatus];


export const ItemStatus = {
  open: 'open',
  resolved: 'resolved',
} as const;

export interface Item {
  id: number;
  type: ItemType;
  category: string;
  color: string;
  brand: string;
  description: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  locationDescription?: string | null;
  status: ItemStatus;
  reporterName: string;
  reporterContact: string;
  createdAt: string;
  updatedAt: string;
}

export type ItemInputType = typeof ItemInputType[keyof typeof ItemInputType];


export const ItemInputType = {
  lost: 'lost',
  found: 'found',
} as const;

export interface ItemInput {
  type: ItemInputType;
  category: string;
  color: string;
  brand: string;
  description: string;
  imageUrl?: string;
  locationDescription?: string;
  reporterName: string;
  reporterContact: string;
}

export type ItemUpdateStatus = typeof ItemUpdateStatus[keyof typeof ItemUpdateStatus];


export const ItemUpdateStatus = {
  open: 'open',
  resolved: 'resolved',
} as const;

export interface ItemUpdate {
  category?: string;
  color?: string;
  brand?: string;
  description?: string;
  imageUrl?: string;
  locationDescription?: string;
  reporterName?: string;
  reporterContact?: string;
  status?: ItemUpdateStatus;
}

export interface CategoryCount {
  category: string;
  count: number;
}

export interface ItemStats {
  totalLost: number;
  totalFound: number;
  totalResolved: number;
  totalOpen: number;
  categoryBreakdown: CategoryCount[];
}

export interface ErrorResponse {
  error: string;
}

export type ListItemsParams = {
type?: ListItemsType;
category?: string;
color?: string;
brand?: string;
status?: ListItemsStatus;
};

export type ListItemsType = typeof ListItemsType[keyof typeof ListItemsType];


export const ListItemsType = {
  lost: 'lost',
  found: 'found',
} as const;

export type ListItemsStatus = typeof ListItemsStatus[keyof typeof ListItemsStatus];


export const ListItemsStatus = {
  open: 'open',
  resolved: 'resolved',
} as const;