# lib/api-zod/src/generated/types/item.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Lost & Found API
 * OpenAPI spec version: 0.1.0
 */
import type { ItemStatus } from './itemStatus';
import type { ItemType } from './itemType';

export interface Item {
  id: number;
  type: ItemType;
  category: string;
  color: string;
  brand: string;
  description: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  locationDescription?: string | null;
  status: ItemStatus;
  reporterName: string;
  reporterContact: string;
  createdAt: Date;
  updatedAt: Date;
}