# lib/api-zod/src/generated/types/itemInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Lost & Found API
 * OpenAPI spec version: 0.1.0
 */
import type { ItemInputType } from './itemInputType';

export interface ItemInput {
  type: ItemInputType;
  category: string;
  color: string;
  brand: string;
  description: string;
  imageUrl?: string;
  locationDescription?: string;
  reporterName: string;
  reporterContact: string;
}