# lib/api-zod/src/generated/types/itemStats.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Lost & Found API
 * OpenAPI spec version: 0.1.0
 */
import type { CategoryCount } from './categoryCount';

export interface ItemStats {
  totalLost: number;
  totalFound: number;
  totalResolved: number;
  totalOpen: number;
  categoryBreakdown: CategoryCount[];
}