# lib/api-zod/src/generated/types/itemUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Lost & Found API
 * OpenAPI spec version: 0.1.0
 */
import type { ItemUpdateStatus } from './itemUpdateStatus';

export interface ItemUpdate {
  category?: string;
  color?: string;
  brand?: string;
  description?: string;
  imageUrl?: string;
  locationDescription?: string;
  reporterName?: string;
  reporterContact?: string;
  status?: ItemUpdateStatus;
}